Functions

LLLplus.lll — Function

B,T,Q,R = lll(H,δ=3/4)

Do Lenstra–Lenstra–Lovász lattice reduction of matrix H using optional parameter δ. The output is B, an LLL-reduced basis; T, a unimodular (meaning det(T)=+/-1) transformation matrix such that B= H*T; and finally Q and R which are a QR decomposition of B. So H = B*inv(T) = Q*R*inv(T).

Follows D. Wuebben, et al, "Lattice Reduction - A Survey with Applications in Wireless Communications", IEEE Signal Processing Magazine, Apr 2011. http://www.ant.uni-bremen.de/sixcms/media.php/102/10740/SPM2011Wuebben.pdf

The LLL tecnique was originally described in "Factoring polynomials with rational coefficients" by A. K. Lenstra, H. W. Lenstra Jr. and L. Lovász, Mathematische Annalen 261, 1982. http://ftp.cs.elte.hu/~lovasz/scans/lll.pdf

Examples

julia> H= [1 2; 3 4];B,_ = lll(H); B
2×2 Array{Int64,2}:
 1  -1
 1   1

julia> H= BigFloat.([1.5 2; 3 4]) .+ 2im; B,_= lll(H); B
2×2 Array{Complex{BigFloat},2}:
 0.50+0.0im  0.0+1.0im
  1.0+0.0im  0.0+0.0im

julia> N=500;H = randn(N,N); B,T = lll(H);

LLLplus.cvp — Function

x=cvp(y,R)

Solve the problem argmin_x ||z-Hx|| for integer x using the technique from the paper below, where H=QR and y=Q'*z. The input vector y is of length n, with H of dimension n by n, and the returned vector x of length n.

x=cvp(y,R,infinite=Val(true),Umax=-1,Umax=-Umin,nxMax=Int(ceil(log2(n)*1e6)))

If infinite==Val(true) then we search the (infinite) lattice, otherwise we search integers in [Umin,Umax]. When nxMax is included, it gives a maximum number of steps to take, otherwise a heuristic value is used. Note that cvp does not handle complex numbers.

Follows "Faster Recursions in Sphere Decoding" Arash Ghasemmehdi, Erik Agrell, IEEE Transactions on Information Theory, vol 57, issue 6 , June 2011.

Examples

julia> H=[1 2; 3 4]; Q,R=qr(H); uhat = cvp(Q'*[0,2],R)
2-element Array{Float64,1}:
  2.0
 -1.0

julia> uhat = cvp(Q'*[0,2],R,Val(false),0,100)
2-element Array{Float64,1}:
 1.0
 0.0

julia> n=100;H=randn(n,n);Q,R=qr(H);

julia> u=Int.(rand(0:1e10,n));y=H*u+rand(n)/100;

julia> uhat=cvp(Q'*y,R); sum(abs.(u-uhat))
0.0

julia> n=500;H=randn(n,n);Q,R=qr(H);

julia> u=Int.(rand([-1,1],n));y=H*u+rand(n)/100;

julia> uhat=cvp(Q'*y,R,Val(false),-1,1); sum(abs.(u-uhat))
0.0

LLLplus.svp — Function

x=svp(B)

Find the shortest basis vector b for the lattice formed by the matrix B. This solves the 'shortest vector problem' (SVP). Note that svp does not handle complex numbers.

Follows "Algorithm SHORTESTVECTOR(G)" from "Closest Point Search in Lattices" by Erik Agrell, Thomas Eriksson, Alexander Vardy, and Kenneth Zeger in IEEE Transactions on Information Theory, vol. 48, no. 8, August 2002.

Examples

julia> H=[1 2; 3 4]; svp(H)
2-element Array{Int64,1}:
 -1
 -1

julia> H = [1 0 0 0;   0 1 0 0;   208 175 663 0;     651 479 0  663];

julia> svp(H)
4-element Array{Int64,1}:
  16
 -19
   3
 -11

LLLplus.brun — Function

B, T = brun(H)

Brun's integer-relations alg implemented as a matrix decomposition. Takes as input the matrix H and returns a reduced basis B and T, a unimodular transformation matrix such that B = H*T. Brun reduction is often done with pinv(H) as input to yield B = pinv(H)*T.

See V. Brun, "En generalisation av kjedebrøken I," Skr. Vid ensk. Selsk. Kristiana, Mat. Nat. Klasse, 1919. See https://archive.org/stream/skrifterutgitavv201chri#page/300/mode/2up

Follows code from D. Wuebben, D. Seethaler, J. Jalden, and G. Matz, "Lattice Reduction - A Survey with Applications in Wireless Communications" IEEE Signal Processing Magazine, March 2011

Examples

julia> H=[1 2; 3 4]; B,T=brun(H); T
2×2 Array{Int64,2}:
  3  -1
 -2   1

LLLplus.gauss — Function

B = gauss(H)

Do Gauss/Lagrange reduction on the lattice defined by the two columns of H.

Follows Fig 2.3 of "Lattice Basis Reduction: An Introduction to the LLL Algorithm and Its Applications" by Murray R. Bremner, CRC Press, 2012.

Examples

julia> H = [1 2; 3 3]; B = gauss(H)
2×2 Array{Float64,2}:
 1.0  0.0
 0.0  3.0

LLLplus.seysen — Function

B,T,B_dual,num_it = seysen(H::Array{Td,2}) where Td

Do greedy Seysen lattice reduction on the matrix H, returning B, the reduced lattice basis; T a unimodular matrix that reduces H (i.e. B = H*T); B_dual, dual lattice basis (i.e., B_dual = pinv(B)); and num_it the number of iterations (basis updates). See also lll.

Follows Seysen algorithm in "Lattice Reduction - A Survey with Applications in Wireless Communications" by D. Wuebben, et al, IEEE Signal Processing Magazine, 2011.

Technique originally described in "Simultaneous reduction of a lattice basis and its reciprocal basis" by M. Seysen, Combinatorica, 1993. http://link.springer.com/article/10.1007%2FBF01202355

Examples

julia> H= [1 2; 3 4];B,T = seysen(H); B
2×2 Array{Int64,2}:
 -1  1
  1  1

julia> H= BigFloat.([1.5 2; 3 4]) .+ 2im; B,_= seysen(H); B
2×2 Array{Complex{BigFloat},2}:
 0.0+1.0im  0.50+0.0im
 0.0+0.0im   1.0+0.0im

LLLplus.vblast — Function

W,P,B = vblast(H)

Find a VBLAST decomposition of H such that H = pinv(W)*B*P' or B = W*H*P. Here P is a permutation matrix, B is lower triangular with ones on the diagonal, and W has orthogonal rows.

W,P,B = vblast(H,mu)

If an SNR argument mu is passed in, a regularized ("MMSE") decomposition is done, with the result that W will no longer have orthogonal rows and B is no longer lower triangular.

Examples

julia> H= [1. 2; 3 4];W,_ = vblast(H); W
2×2 Array{Float64,2}:
 1.5  -0.5
 0.1   0.3

julia> H= BigFloat.([1.5 2; 3 4]) .+ 2im; W,_= vblast(H); W
2×2 Array{Complex{BigFloat},2}:
      -2.0+3.0im            2.0-1.5im     
 0.0779221-0.103896im  0.155844-0.103896im

LLLplus.subsetsum — Function

x = subsetsum(a,s)

For a vector of integers a, and an integer s, try to find a binary vector x such that x'*a=s. We use the LLL algorithm to find the solution. This is not a robust tool, just a demo.

This function tries first the technique in the lagariasodlyzko function, and if it fails, a solution via mdsubsetsum is attempted.

It appears that this function can also solve some integer relations problems. See the first example.

Examples

julia> a=[1.5;.5;0;.1;.2]; s=2.2; x,_=subsetsum(a,s,true); s-x'*a
A binary Lagarias-Odlyzko solution was found.
A solution was found via lagariasodlyzko
0.0

julia> a=[32771,65543,131101,262187,524387,1048759, # from Bremner p 117
          2097523,4195057,8390143,16780259,33560539,
          67121039,134242091,268484171,536968403];

julia> s=891221976; x,_=subsetsum(a,s,false); s-x'*a
0.0

julia> N=40;a=rand(1:2^BigInt(256),N);xtrue=rand(Bool,N); s=a'*xtrue; 

julia> setprecision(BigFloat,300); x,_=subsetsum(a,s,false); s-x'*a
0.0

LLLplus.lagariasodlyzko — Function

x = lagariasodlyzko(a,s)

For a vector of integers a, and an integer s, try to find a binary vector x such that x'*a=s. We use the LLL algorithm to find the solution. This is not a robust tool, just a demo.

This follows the technique described by Lagarias and Odlyzko in "Solving Low-Density Subset Sum Problems" in Journal of ACM, Jan 1985. Code based on http://web.eecs.umich.edu/~cpeikert/lic15/lec05.pdf We can likely get better results using techniques described and referenced in https://www-almasty.lip6.fr/~joux/pages/papers/ToolBox.pdf

It's odd that permuting the a vector in the second example given below causes the alg to often not find a binary solution. Apparently this is a common oddity with lattice solvers.

Examples

julia> a=[1.5;.5;0;.1;.2]; s=2.2; x,_=lagariasodlyzko(a,s); s-x'*a
0.0

julia> a=[32771,65543,131101,262187,524387,1048759, # from Bremner p 117
          2097523,4195057,8390143,16780259,33560539,
          67121039,134242091,268484171,536968403];

julia> s=891221976; x,_=lagariasodlyzko(a,s); s-x'*a
0.0

julia> N=40;a=rand(1:2^BigInt(256),N);xtrue=rand(Bool,N); s=a'*xtrue; 

julia> setprecision(BigFloat,300); x,_=lagariasodlyzko(a,s); s-x'*a
0.0

LLLplus.mdsubsetsum — Function

x = mdsubsetsum(a,sM,ratio=.5,Kpm=3)

For a vector of integers a, and an integer sM, try to find a binary vector x such that x'*a=s using the technique from "Multidimensional subset sum problem" [1][2]. A major goal of the technique is to solve problems in which there are about 50% ones in x; other ratios of ones to zeros can be specified in ratio. The thesis also suggests searching Kpm=3 values around the nominal k. This technique is related to that in subsetsum in that both use the LLL algorithm. This is not a robust tool, just a demo.

[1] https://scholarworks.rit.edu/theses/64/ [2] https://pdfs.semanticscholar.org/21a7/c2f9ff29507f1153aefcca04d1cd308e45c0.pdf

Examples

julia> a=[1.5;.5;0;.1;.2]; s=2.2; x=mdsubsetsum(a,s); s-x'*a
0.0

julia> a=[32771,65543,131101,262187,524387,1048759, # from Bremner p 117
          2097523,4195057,8390143,16780259,33560539,
          67121039,134242091,268484171,536968403];

julia> sM=891221976; x=mdsubsetsum(a,sM); sM-x'*a
0

julia> setprecision(BigFloat,300); Random.seed!(0);

julia> N=40;a=rand(1:2^BigInt(256),N);xtrue=rand(Bool,N); s=a'*xtrue;

julia> x=mdsubsetsum(a,s); s-x'*a
0

LLLplus.integerfeasibility — Function

integerfeasibility(A,d,nullVecs=false)

Given a linear system Ax=d, return an integer vector x which satisfies the system. This is the integer programming feasibility problem.

If nullVecs==true, then as well as returning a solution x, also return a matrix xNull of vectors in the null space of A which could be added to the x vector to find a solution which satisfies a constraint such as 0 .≤ x .≤ u; see the paper below.

This is not a robust tool, just a demo.

"Solving A System Of Diophantine Equations With Bounds On The Variables" by Karen Aardal, Cor Hurkens, and Arjen Lenstra in Integer Programming and Combinatorial Optimization, 6th International IPCO Conference, vol 1412, pp 229-242, 1998. See http://softlib.rice.edu/pub/CRPC-TRs/reports/CRPC-TR98782.pdf

Examples

julia> A=[10 1 -9; 1 8 8]; xtrue=[0; 2; 9]; d=A*xtrue;

julia> integerfeasibility(A,d)
3-element Array{Int64,1}:
 0
 2
 9

julia> A=[10 1.1 -9.1; 1 8 8]; d=A*xtrue;

julia> integerfeasibility(A,d)
3-element Array{Float64,1}:
 0.0
 2.0
 9.0

julia> n=20;m=30; A = rand(-10:10,n,m); xtrue = rand(0:10,m); d=A*xtrue;

julia> sum(abs.(xtrue - integerfeasibility(A,d) ))
0

LLLplus.rationalapprox — Function

rationalapprox(x::AbstractArray{<:Real,1},M,Ti=BigInt,verbose=false)

For a vector of Reals x, and an integer M, find an integer q such that maximum(abs.(x*q-round.(x*q))) is small; the vector x is approximated by round.(x*q)//q. The integer q is less than or equal to M and the approximation satisfies max(abs.(x*q-round.(x*q)))≤sqrt(5)*2^(n/4 - 5)*M^(-1/n); this equation comes from the paper below. The LLL algorithm reduction is used to find the solution. The approximation vector is returned. This is also known as "simultaneous diophantine approximation"; see for example the title of the Hanrot paper below.

This is not a robust tool, just a demo.

"LLL: A Tool for Effective Diophantine Approximation" by Guillaume Hanrot in the book "The LLL Algorithm: Survey and Applications" edited by Phong Q. Nguyen and Brigitte Vallée, Springer, Heidelberg, 2010.

See also Chapter 9 of M. R. Bremner, "Lattice Basis Reduction: An Introduction to the LLL Algorithm and Its Applications" CRC Press, 2012.

Examples

julia> x = [0.3912641745333527; 0.5455179974014548; 0.1908698210882469];

julia> rationalapprox(x,1e4,Int64)
3-element Array{Rational{Int64},1}:
 43//110
  6//11
 21//110

LLLplus.spigotBBP — Function

spigotBBP(α::Td,s,b,n,K,verbose=false) where {Td}

Check for a BBP-style [1] infinite series for the constant α. These are "spigot" formulas that can be used to generate (for example) the millionth digit of the constant α without learning the previous digits. Specifically, given the constant α, and parameters b, n, and s, look for a vector of numbers a_1 through a_n that satisfies the following equation:

$\alpha= \sum_{k=0}^\infty \frac{1}{b^k} \left( \frac{a_1}{(nk+1)^s} + \ldots + \frac{a_n}{(nk+n)^s} \right)$

Because it's hard to sum to infinity, the sum is stopped at K. If a formula is found, it is printed to the screen in LaTeX and the coefficents a are returned as a vector. An online LaTeX viewer such as https://www.latex4technics.com/ may be helpful.

This is not a robust tool, just a demo. For example, there may be a problem with s≥2. See [2] for derivation of the technique used, and to check whether a formula you find is new.

[1] David Bailey, Peter Borwein, and Simon Plouffe. "On the rapid computation of various polylogarithmic constants." Mathematics of Computation 66.218 (1997): 903-913. https://www.ams.org/journals/mcom/1997-66-218/S0025-5718-97-00856-9/

[2] David Bailey, "A Compendium of BBP-Type Formulas for Mathematical Constants". https://www.davidhbailey.com//dhbpapers/bbp-formulas.pdf

Example

julia> spigotBBP(BigFloat(pi),1,16,8,45,true);
A solution was found w error -4.728672e-60. In LaTeX form it is
\alpha= \sum_{k=0}^\infty \frac{1}{16^k} \left(\frac{4}{8k+1}-\frac{2}{8k+4}-\frac{1}{8k+5}-\frac{1}{8k+6}\right)

Other examples without output:

spigotBBP(Float64(pi),1,-4,4,22,true);
spigotBBP(log(2),1,2,2,30,true);
spigotBBP(9*log(3),1,9,2,30,true);
spigotBBP(atan(2)*8,1,16,8,30,true);
spigotBBP(8*sqrt(2)*log(1+sqrt(2)),1,16,8,25,true);

There is a formula for pi^2 which the following command should find, but it does not find it. In fact the technique doesn't seem to work at all for s>2; It's not obvious what the problem is

spigotBBP(BigFloat(pi)*pi,2,64,6,25,true);

LLLplus.hard_sphere — Function

hard_sphere(...) = hardsphere(...)

See hardsphere; in a future version hard_sphere will be removed.

LLLplus.issizereduced — Function

issizereduced(B)

Determine if the matrix B is size reduced or not.

Examples

julia> H= [1 2; 3 4];issizereduced(H)
false

julia> B,T = lll(H);issizereduced(B)
true

LLLplus.islllreduced — Function

islllreduced(B)

Determine if the matrix B is LLL reduced or not. See p 56 of Bremner for a definition.

M. R. Bremner, "Lattice Basis Reduction: An Introduction to the LLL Algorithm and Its Applications" CRC Press, 2012.

Examples

julia> H= [1 2; 3 4];islllreduced(H)
false

julia> B,T=lll(H);islllreduced(B)
true

LLLplus.orthogonalitydefect — Function

orthogonalitydefect(B)

Find the orthogonality defect of the matrix B defined, for example, on page 2 of Bennet:

"An Enumeration Technique for Lattice Basis Reduction" Huck Bennet, 2 Nov

https://web.eecs.umich.edu/~hdbco/publications/bases.pdf

Examples

julia> H= [1 2; 3 4];B,T=lll(H);

julia> [orthogonalitydefect(H) orthogonalitydefect(B)]
1×2 Array{Float64,2}:
 7.07107  1.0

LLLplus.hermitefactor — Function

hermitefactor(B)

Find the Hermite factor of matrix B

Examples

julia> H= [1 2; 3 4];hermitefactor(H)
1.5811388300841898

LLLplus.seysencond — Function

seysencond(B)

Seysen condition number as on, for example, page 3 of Bennet

Bennet

Examples

julia> H= [1 2; 3 4];seysencond(H)
2.8284271247461903

LLLplus.gen_qary_b — Function

b= gen_qary_b(T, d::Int,k::Int,b::Int)

Generate a q-ary lattice given an element type T, dimension d, parameter k, and bit-depth b. Specifically, find a d by d matrix which has the block structure [I zeros(T,k,d-k); H q*I]], where the k by d-k matrix H is sampled from 0:q-1 and q is sampled uniformly from 1:big(2)^b-1

These bases correspond to the SIS/LWE q-ary lattices; see D. Micciancio and O. Regev. Post-Quantum Cryptography. Chapter of Lattice-based Cryptography, 147-191 (2009) and latticegen in https://github.com/fplll/fplll

Examples

julia> b=gen_qary_b(Int64,2,1,6)
2×2 Array{Int64,2}:
 1   0
 7  32